# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 1WIN Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResult {
  error: string;
}

export interface ActionResult {
  success: boolean;
  message: string;
}

export interface LoginInput {
  password: string;
}

export interface LoginResult {
  token: string;
}

export interface Stats {
  totalUsers: number;
  approvedUsers: number;
  premiumUsers: number;
  pendingUsers: number;
  bannedUsers: number;
  activeToday: number;
  totalSignals: number;
}

export interface User {
  telegramId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  languageCode?: string;
  isPremium: boolean;
  isBanned: boolean;
  hasRegistered: boolean;
  approvalStatus: string;
  totalAnalyses: number;
  freeSignalsUsedToday?: number;
  /** @nullable */
  registeredAt: string | null;
  /** @nullable */
  lastActive?: string | null;
}

export interface UserList {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface PremiumInput {
  active: boolean;
  days: number;
}

export interface BotSettings {
  promoCode: string;
  affiliateLink: string;
  freeSignalsPerDay: number;
  premiumSignalsPerDay: number;
  channel1Link: string;
  channel1Name: string;
  channel2Link: string;
  channel2Name: string;
  /** @nullable */
  menuBanner?: string | null;
  /** @nullable */
  registerBanner?: string | null;
  /** @nullable */
  luckyjetBanner?: string | null;
  /** @nullable */
  minesBanner?: string | null;
  /** @nullable */
  guideBanner?: string | null;
}

export interface BotSettingsInput {
  promoCode?: string;
  affiliateLink?: string;
  freeSignalsPerDay?: number;
  premiumSignalsPerDay?: number;
  channel1Link?: string;
  channel1Name?: string;
  channel2Link?: string;
  channel2Name?: string;
}

export interface BroadcastInput {
  message: string;
}

export interface BroadcastResult {
  sent: number;
  failed: number;
}

export type ListUsersParams = {
status?: ListUsersStatus;
page?: number;
limit?: number;
search?: string;
};

export type ListUsersStatus = typeof ListUsersStatus[keyof typeof ListUsersStatus];


export const ListUsersStatus = {
  all: 'all',
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
  banned: 'banned',
} as const;