# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 1WIN Bot Admin Panel API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: stats
    description: Dashboard statistics
  - name: users
    description: User management
  - name: settings
    description: Bot settings
  - name: broadcast
    description: Broadcast messaging

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Login with admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /admin/stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get dashboard statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /admin/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users with optional status filter
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, rejected, banned]
          required: false
        - name: page
          in: query
          schema:
            type: integer
          required: false
        - name: limit
          in: query
          schema:
            type: integer
          required: false
        - name: search
          in: query
          schema:
            type: string
          required: false
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserList"

  /admin/users/{telegramId}/approve:
    post:
      operationId: approveUser
      tags: [users]
      summary: Approve a pending user
      security:
        - bearerAuth: []
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResult"

  /admin/users/{telegramId}/reject:
    post:
      operationId: rejectUser
      tags: [users]
      summary: Reject a user
      security:
        - bearerAuth: []
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /admin/users/{telegramId}/ban:
    post:
      operationId: banUser
      tags: [users]
      summary: Ban a user
      security:
        - bearerAuth: []
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /admin/users/{telegramId}/unban:
    post:
      operationId: unbanUser
      tags: [users]
      summary: Unban a user
      security:
        - bearerAuth: []
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User unbanned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /admin/users/{telegramId}/premium:
    post:
      operationId: setPremium
      tags: [users]
      summary: Toggle premium for a user
      security:
        - bearerAuth: []
      parameters:
        - name: telegramId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PremiumInput"
      responses:
        "200":
          description: Premium updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResult"

  /admin/settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all bot settings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bot settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update bot settings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotSettingsInput"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotSettings"

  /admin/broadcast:
    post:
      operationId: sendBroadcast
      tags: [broadcast]
      summary: Send broadcast to all approved users
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status:
          type: string

    ErrorResult:
      type: object
      required: [error]
      properties:
        error:
          type: string

    ActionResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string

    LoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    LoginResult:
      type: object
      required: [token]
      properties:
        token:
          type: string

    Stats:
      type: object
      required: [totalUsers, premiumUsers, pendingUsers, bannedUsers, activeToday, totalSignals, approvedUsers]
      properties:
        totalUsers:
          type: integer
        approvedUsers:
          type: integer
        premiumUsers:
          type: integer
        pendingUsers:
          type: integer
        bannedUsers:
          type: integer
        activeToday:
          type: integer
        totalSignals:
          type: integer

    User:
      type: object
      required: [telegramId, approvalStatus, isPremium, isBanned, hasRegistered, totalAnalyses, registeredAt]
      properties:
        telegramId:
          type: integer
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        languageCode:
          type: string
        isPremium:
          type: boolean
        isBanned:
          type: boolean
        hasRegistered:
          type: boolean
        approvalStatus:
          type: string
        totalAnalyses:
          type: integer
        freeSignalsUsedToday:
          type: integer
        registeredAt:
          type: ["string", "null"]
        lastActive:
          type: ["string", "null"]

    UserList:
      type: object
      required: [users, total, page, limit]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    PremiumInput:
      type: object
      required: [active, days]
      properties:
        active:
          type: boolean
        days:
          type: integer

    BotSettings:
      type: object
      required: [promoCode, affiliateLink, freeSignalsPerDay, premiumSignalsPerDay, channel1Link, channel1Name, channel2Link, channel2Name]
      properties:
        promoCode:
          type: string
        affiliateLink:
          type: string
        freeSignalsPerDay:
          type: integer
        premiumSignalsPerDay:
          type: integer
        channel1Link:
          type: string
        channel1Name:
          type: string
        channel2Link:
          type: string
        channel2Name:
          type: string
        menuBanner:
          type: ["string", "null"]
        registerBanner:
          type: ["string", "null"]
        luckyjetBanner:
          type: ["string", "null"]
        minesBanner:
          type: ["string", "null"]
        guideBanner:
          type: ["string", "null"]

    BotSettingsInput:
      type: object
      properties:
        promoCode:
          type: string
        affiliateLink:
          type: string
        freeSignalsPerDay:
          type: integer
        premiumSignalsPerDay:
          type: integer
        channel1Link:
          type: string
        channel1Name:
          type: string
        channel2Link:
          type: string
        channel2Name:
          type: string

    BroadcastInput:
      type: object
      required: [message]
      properties:
        message:
          type: string

    BroadcastResult:
      type: object
      required: [sent, failed]
      properties:
        sent:
          type: integer
        failed:
          type: integer