# lib/api-zod/src/generated/types/botSettings.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 1WIN Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface BotSettings {
  promoCode: string;
  affiliateLink: string;
  freeSignalsPerDay: number;
  premiumSignalsPerDay: number;
  channel1Link: string;
  channel1Name: string;
  channel2Link: string;
  channel2Name: string;
  /** @nullable */
  menuBanner?: string | null;
  /** @nullable */
  registerBanner?: string | null;
  /** @nullable */
  luckyjetBanner?: string | null;
  /** @nullable */
  minesBanner?: string | null;
  /** @nullable */
  guideBanner?: string | null;
}