# lib/api-zod/src/generated/types/premiumInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 1WIN Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface PremiumInput {
  active: boolean;
  days: number;
}