# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * 1WIN Bot Admin Panel API
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  telegramId: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  firstName?: string | null;
  /** @nullable */
  lastName?: string | null;
  languageCode?: string;
  isPremium: boolean;
  isBanned: boolean;
  hasRegistered: boolean;
  approvalStatus: string;
  totalAnalyses: number;
  freeSignalsUsedToday?: number;
  /** @nullable */
  registeredAt: string | null;
  /** @nullable */
  lastActive?: string | null;
}